feat: Coordinator has no knowledge about the master type
The Coordinator is now agnostic to the master child class. Note that for
now, it is not completely perfect since we still have knowledge of the
arguments to build the Master object, e.g. fixed_point_init.
But this is a first step.

The config parser now builds a config dict for Master inspired from the
stream handler config:
- a "type" key, defining which master child class to instantiate.
- a "config" key, containing all the arguments used to instantiate the
  master algorithm.